first version of 2q was implemented and tested

+++ core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/OStorageLocal.java
@@ -58,8 +58,8 @@ import com.orientechnologies.orient.core.id.OClusterPosition;
 import com.orientechnologies.orient.core.id.OClusterPositionFactory;
 import com.orientechnologies.orient.core.id.ORID;
 import com.orientechnologies.orient.core.id.ORecordId;
+import com.orientechnologies.orient.core.index.hashindex.local.arc.O2QCache;
 import com.orientechnologies.orient.core.index.hashindex.local.cache.ODiskCache;
-import com.orientechnologies.orient.core.index.hashindex.local.cache.OLRUCache;
 import com.orientechnologies.orient.core.memory.OMemoryWatchDog;
 import com.orientechnologies.orient.core.metadata.OMetadata;
 import com.orientechnologies.orient.core.storage.OCluster;
@@ -131,7 +131,7 @@ public class OStorageLocal extends OStorageEmbedded {
     final ODirectMemory directMemory = ODirectMemoryFactory.INSTANCE.directMemory();
 
     if (directMemory != null)
-      diskCache = new OLRUCache(OGlobalConfiguration.DISK_CACHE_SIZE.getValueAsLong() * 1024 * 1024, directMemory,
+      diskCache = new O2QCache(OGlobalConfiguration.DISK_CACHE_SIZE.getValueAsLong() * 1024 * 1024, directMemory,
           OGlobalConfiguration.DISK_CACHE_PAGE_SIZE.getValueAsInteger(), this, false);
     else
       diskCache = null;
@@ -418,15 +418,18 @@ public class OStorageLocal extends OStorageEmbedded {
           int notDeletedFiles = 0;
 
           // TRY TO DELETE ALL THE FILES
-          for (File f : dbDir.listFiles()) {
-            // DELETE ONLY THE SUPPORTED FILES
-            for (String ext : ALL_FILE_EXTENSIONS)
-              if (f.getPath().endsWith(ext)) {
-                if (!f.delete()) {
-                  notDeletedFiles++;
+          File[] files = dbDir.listFiles();
+          if (files != null) {
+            for (File f : files) {
+              // DELETE ONLY THE SUPPORTED FILES
+              for (String ext : ALL_FILE_EXTENSIONS)
+                if (f.getPath().endsWith(ext)) {
+                  if (!f.delete()) {
+                    notDeletedFiles++;
+                  }
+                  break;
                 }
-                break;
-              }
+            }
           }
 
           if (notDeletedFiles == 0) {
@@ -588,7 +591,8 @@ public class OStorageLocal extends OStorageEmbedded {
         // CHECK CHUNKS
         formatMessage(iVerbose, iListener, "\n-- checking chunks:");
 
-        for (int pos = 0; nextPos < d.getFilledUpTo();) {
+        int pos;
+        do {
           try {
             pos = nextPos;
 
@@ -704,7 +708,7 @@ public class OStorageLocal extends OStorageEmbedded {
             // totalChunks, pos, e.toString());
             errors++;
           }
-        }
+        } while (nextPos < d.getFilledUpTo());
         formatMessage(iVerbose, iListener, "\n");
       }
 
@@ -1350,8 +1354,7 @@ public class OStorageLocal extends OStorageEmbedded {
 
   /**
    * Returns the list of holes as pair of position & ODataHoleInfo
-   * 
-   * @throws IOException
+   *
    */
   public List<ODataHoleInfo> getHolesList() {
     final List<ODataHoleInfo> holes = new ArrayList<ODataHoleInfo>();
@@ -1373,7 +1376,6 @@ public class OStorageLocal extends OStorageEmbedded {
   /**
    * Returns the total number of holes.
    * 
-   * @throws IOException
    */
   public long getHoles() {
     lock.acquireSharedLock();
@@ -1393,7 +1395,6 @@ public class OStorageLocal extends OStorageEmbedded {
   /**
    * Returns the total size used by holes
    * 
-   * @throws IOException
    */
   public long getHoleSize() {
     lock.acquireSharedLock();
@@ -1591,7 +1592,6 @@ public class OStorageLocal extends OStorageEmbedded {
    * @param iConfig
    *          A OStorageClusterConfiguration implementation, namely physical or logical
    * @return The id (physical position into the array) of the new cluster just created. First is 0.
-   * @throws IOException
    * @throws IOException
    */
   private int createClusterFromConfig(final OStorageClusterConfiguration iConfig) throws IOException {
